+++ prime_sdk/services/financing/set_funding_settings.py
@@ -15,28 +15,21 @@
 from dataclasses import dataclass
 from typing import List, Optional
 from ...base_response import BaseResponse
-from ...enums import OrderSide
 
 
 @dataclass
-class CreateQuoteRequest:
-    portfolio_id: str
-    product_id: str
-    side: OrderSide
-    client_quote_id: str
-    limit_price: str
-    base_quantity: Optional[str] = None
-    quote_value: Optional[str] = None
-    settl_currency: Optional[str] = None
-    quote_duration_ms: Optional[str] = None
+class SetFundingSettingsRequest:
+    entity_id: str
+    designated_funding_portfolio_id: str
+    automatic_conversion_enabled: bool
+    automatic_loan_enabled: bool
+    automatic_excess_return_enabled: bool
+    excess_funds_target_amount: str
     allowed_status_codes: Optional[List[int]] = None
 
 
 @dataclass
-class CreateQuoteResponse(BaseResponse):
-    quote_id: str = None
-    expiration_time: str = None
-    best_price: str = None
-    order_total: str = None
-    price_inclusive_of_fees: str = None
-    quote_duration_ms: str = None
+class SetFundingSettingsResponse(BaseResponse):
+    activity_id: str = None
+    activity_type: str = None
+    num_approvals_remaining: int = None